Applies manual versioning to static assets.
This allows us to invalidate browser caches when the need arises;
particular files may also be excluded when necessary, and in this
case semantic themes (fonts) are exempted from the versioning requirement.

As a consequence, I also corrected all the asset url's in the HTML files
so they request the assets under the resources/ folder.

The favicon also works now.