Replace semantic imports with jsdelivr links (check desc).
Note that only these links were replaced these assets are not liable to change
often. I am not sure how often jsdelivr caches it's files, but I doubt
it's something we can modify. For the other assets, we can at least
specify a ?v= parameter to help us in that regard.